backend/requirements.txt
alembic==1.0.8
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-cloud-videointelligence==1.8.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.2
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
